Move reusable test functions to testsuite crate
