+++ testsuite/src/lib.rs
@@ -10,7 +10,7 @@
 use defmt_rtt as _;
 use panic_probe as _;
 
-use bxcan::{FilterOwner, Instance, MasterInstance};
+use bxcan::{Can, FilterOwner, Frame, Instance, MasterInstance};
 
 pub use stm32f1::stm32f107 as pac;
 
@@ -37,7 +37,7 @@ unsafe impl Instance for CAN2 {
     const REGISTERS: *mut bxcan::RegisterBlock = 0x4000_6800 as *mut _;
 }
 
-pub fn init(p: pac::Peripherals) -> (CAN1, CAN2) {
+fn init(p: pac::Peripherals) -> (CAN1, CAN2) {
     p.RCC
         .apb1enr
         .modify(|_, w| w.can1en().enabled().can2en().enabled());
@@ -67,4 +67,72 @@ pub fn init(p: pac::Peripherals) -> (CAN1, CAN2) {
     let _ = (p.CAN1, p.CAN2);
 
     (CAN1 { _private: () }, CAN2 { _private: () })
+}
+
+pub struct State {
+    pub can1: Can<CAN1>,
+    pub can2: Can<CAN2>,
+}
+
+impl State {
+    pub fn init() -> Self {
+        let periph = defmt::unwrap!(pac::Peripherals::take());
+        let (can1, can2) = init(periph);
+        let mut can1 = Can::new(can1);
+        let can2 = Can::new(can2);
+        can1.modify_filters().clear();
+
+        let mut state = Self { can1, can2 };
+        state.go_fast();
+        state
+    }
+
+    /// Configures the slowest possible speed.
+    ///
+    /// This is useful for testing recovery when the mailboxes are full.
+    pub fn go_slow(&mut self) {
+        self.can1
+            .modify_config()
+            .set_loopback(true)
+            .set_silent(true)
+            .set_bit_timing(0x007f_03ff);
+        self.can2
+            .modify_config()
+            .set_loopback(true)
+            .set_silent(true)
+            .set_bit_timing(0x007f_03ff);
+        nb::block!(self.can1.enable()).unwrap();
+    }
+
+    /// Configures the default (fast) speed.
+    pub fn go_fast(&mut self) {
+        self.can1
+            .modify_config()
+            .set_loopback(true)
+            .set_silent(true)
+            .set_bit_timing(0x00050000);
+        self.can2
+            .modify_config()
+            .set_loopback(true)
+            .set_silent(true)
+            .set_bit_timing(0x00050000);
+        nb::block!(self.can1.enable()).unwrap();
+    }
+
+    pub fn roundtrip_frame(&mut self, frame: &Frame) -> bool {
+        nb::block!(self.can1.transmit(frame)).unwrap();
+        defmt::assert!(!self.can1.is_transmitter_idle());
+
+        // Wait until the transmission has completed.
+        while !self.can1.is_transmitter_idle() {}
+
+        match self.can1.receive() {
+            Ok(received) => {
+                defmt::assert_eq!(received, *frame);
+                true
+            }
+            Err(nb::Error::WouldBlock) => false,
+            Err(nb::Error::Other(e)) => defmt::panic!("{:?}", e),
+        }
+    }
 }
